17/12 1008

+++ app/src/main/java/com/example/fyp/ProfileSetupActivity.kt
@@ -39,7 +39,7 @@ class ProfileSetupActivity : AppCompatActivity() {
             val contactStr = userContact.text.toString()
             val gender = userGender.selectedItem.toString()
             val userId = currentUser?.uid
-            val contact = contactStr.toInt()
+            val contact = contactStr
 
             if (userId != null) {
                 val user = User(
